feat: move all model definitions in model.go
